implement BaseTest
