+++ ApplePieUITests/ApplePieUITests.swift
@@ -7,27 +7,7 @@
 
 import XCTest
 
-class ApplePieUITests: XCTestCase {
-    
-    let app = XCUIApplication()
-    
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-        app.launch()
-        // In UI tests it is usually best to stop immediately when a failure occurs.
-        continueAfterFailure = false
-        // In UI tests it’s important to set the initial state - such as interface orientation - required for your tests before they run. The setUp method is a good place to do this.
-    }
-    
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-    func enterWord(word: String) {
-        for letter in word {
-            let button = app.buttons[String(letter).uppercased()]
-            button.tap()
-        }
-    }
+class ApplePieUITests: BaseTest {
     
     func testStatisticUpdatedAfterGuess() throws {
         enterWord(word: "buccaneer")
